+++ src/widgets/subs.ts
@@ -1,7 +1,7 @@
 import { Subscription } from '../types/subscription';
 import { Config } from '../types/config';
 import { WidgetData } from '../types/widgetData';
-import { getNormalizedValues, startTask, stopTask } from '../app/utils';
+import { getNormalizedValues, startTask, stopTask } from '../utils';
 
 const io = require("socket.io/client-dist/socket.io.min"); // use socket.io/client-dist instead of socket.io-client because streamlabs requires older client version
